Handle playing-state broadcasts on clients that are still loading
The server flips a joiner to the playing state as soon as the world
data is queued, without waiting for the client to finish downloading
and loading it, so player-list, chat and notification broadcasts can
legally arrive while the client is still in the loading state - and
Steam P2P delivers reliable packets out of order besides (a player-list
packet was observed arriving ahead of the world-data header). An
unhandled reliable packet is treated as protocol corruption and kills
the whole session, dumping the joiner back to the menu mid-join.

The three handlers move to the shared client state base, where
keep-alive and time-control handling already live, so loading clients
process them instead of dying. All three touch only session UI state,
and the full player list sent after the world data reconciles anything
received during the window. Sim-relevant packets (commands, freezes,
desync checks) intentionally stay playing-only.
