+++ Source/Client/Networking/State/ClientPlayingState.cs
@@ -24,77 +24,6 @@ namespace Multiplayer.Client
         }
 
         [TypedPacketHandler]
-        public void HandlePlayerList(ServerPlayerListPacket packet)
-        {
-            if (packet.action == PlayerListAction.Add)
-            {
-                foreach (var info in packet.players)
-                {
-                    if (!Multiplayer.session.players.Any(p => p.id == info.id || p.username == info.username))
-                    {
-                        ServerLog.Log($"PlayerList: Adding player {info.id}:{info.username}");
-                        Multiplayer.session.players.Add(PlayerInfo.FromNet(info));
-                    }
-                    else
-                    {
-                        ServerLog.Error($"PlayerList: Adding player {info.id}:{info.username} - player already exists");
-                    }
-                }
-            }
-            else if (packet.action == PlayerListAction.Remove)
-            {
-                ServerLog.Log($"PlayerList: Removing player with id {packet.playerId}");
-                var matches = Multiplayer.session.players.RemoveAll(p => p.id == packet.playerId);
-                if (matches > 1)
-                {
-                    ServerLog.Error($"PlayerList: Removing player with id {packet.playerId} -- occurred {matches} times. This should not happen");
-                }
-            }
-            else if (packet.action == PlayerListAction.List)
-            {
-                ServerLog.Log($"PlayerList: Received player list with {packet.players.Length} entries");
-
-                Multiplayer.session.players.Clear();
-                foreach (var info in packet.players)
-                {
-                    ServerLog.Log($"PlayerList: Adding player from list {info.id}:{info.username}");
-                    Multiplayer.session.players.Add(PlayerInfo.FromNet(info));
-                }
-            }
-            else if (packet.action == PlayerListAction.Latencies)
-            {
-                foreach (var latency in packet.latencies)
-                {
-                    var player = Multiplayer.session.GetPlayerInfo(latency.playerId);
-                    if (player == null)
-                    {
-                        ServerLog.Log($"PlayerList: Received latency info for unknown player with id {latency.playerId}");
-                        continue;
-                    }
-                    player.latency = latency.latency;
-                    player.ticksBehind = latency.ticksBehind;
-                    player.simulating = latency.simulating;
-                    player.frameTime = latency.frameTime;
-                }
-            }
-            else if (packet.action == PlayerListAction.Status)
-            {
-                var player = Multiplayer.session.GetPlayerInfo(packet.playerId);
-                if (player == null)
-                {
-                    ServerLog.Log($"PlayerList: Received player status ({packet.status}) for unknown player with id {packet.playerId}");
-                }
-                else
-                {
-                    player.status = packet.status;
-                }
-            }
-        }
-
-        [TypedPacketHandler]
-        public void HandleChat(ServerChatPacket packet) => Multiplayer.session.AddMsg(packet.msg, rawMessage: packet.rawMessage);
-
-        [TypedPacketHandler]
         public void HandleCursor(ServerCursorPacket packet)
         {
             var player = Multiplayer.session.GetPlayerInfo(packet.playerId);
@@ -155,15 +84,6 @@ namespace Multiplayer.Client
                 var mapsToLoad = Find.Maps.Select(m => m.uniqueID).Append(mapId).Distinct().ToList();
                 Loader.ReloadGame(mapsToLoad, false, Multiplayer.game?.gameComp.asyncTime ?? false);
             });
-        }
-
-        [TypedPacketHandler]
-        public void HandleNotification(ServerNotificationPacket packet)
-        {
-            var namedArgs = Array.ConvertAll(packet.args, s => (NamedArgument)s);
-            var msg = packet.key.Translate(namedArgs);
-            Messages.Message(msg, MessageTypeDefOf.SilentInput, false);
-            ServerLog.Log($"Notification: {msg} ({packet.key}, {packet.args.Join(", ")})");
         }
 
         [TypedPacketHandler]
